# docs/swagger.yaml
definitions:
  models.CashierSalesReport:
    properties:
      cashier:
        type: string
      products:
        items:
          $ref: '#/definitions/models.ProductReport'
        type: array
      sales_count:
        type: integer
      total_amount:
        type: number
      total_items:
        type: integer
    type: object
...
  models.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  models.ProductReport:
    properties:
      product_name:
        type: string
      total_amount:
        type: number
      total_quantity:
        type: integer
    type: object
...
  models.Sale:
    properties:
      created_at:
        type: string
      id:
        type: integer
      total_sum:
        type: number
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  models.SaleItem:
    properties:
      created_at:
        type: string
      id:
        type: integer
      price:
        type: number
      product_id:
        type: integer
      quantity:
        type: integer
      sale_id:
        type: integer
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /reports/sales:
    get:
      description: Возвращает отчёт по продажам между датами from и to (в формате
        YYYY-MM-DD)
      parameters:
      - description: Дата начала периода (YYYY-MM-DD)
        in: query
        name: from
        required: true
        type: string
      - description: Дата конца периода (YYYY-MM-DD)
        in: query
        name: to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CashierSalesReport'
        "400":
          description: Параметры запроса from и to обязательны
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка сервера при формировании отчёта
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Получить отчёт по продажам за период
      tags:
      - reports
  /sale-items:
    get:
      description: Возвращает список всех позиций продажи (SaleItems)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SaleItem'
            type: array
        "500":
          description: Ошибка сервера при получении позиций продажи
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Получить все позиции продажи
      tags:
      - sale-items
    post:
      consumes:
      - application/json
      description: Создает новую позицию продажи
      parameters:
      - description: Данные позиции продажи
        in: body
        name: saleItem
        required: true
        schema:
          $ref: '#/definitions/models.SaleItem'
      produces:
      - application/json
      responses:
        "200":
          description: Позиция продажи успешно создана
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка сервера при создании позиции продажи
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Создать позицию продажи
      tags:
      - sale-items
  /sale-items/{id}:
    delete:
      description: Удаляет позицию продажи по ID
      parameters:
      - description: ID позиции продажи
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Позиция продажи успешно удалена
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверный ID позиции продажи
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка сервера при удалении позиции продажи
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Удалить позицию продажи
      tags:
      - sale-items
    get:
      description: Возвращает позицию продажи по идентификатору
      parameters:
      - description: ID позиции продажи
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SaleItem'
        "400":
          description: Неверный ID позиции продажи
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка сервера при получении позиции продажи
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Получить позицию продажи по ID
      tags:
      - sale-items
    put:
      consumes:
      - application/json
      description: Обновляет данные позиции продажи по ID
      parameters:
      - description: ID позиции продажи
        in: path
        name: id
        required: true
        type: integer
      - description: Обновленные данные позиции продажи
        in: body
        name: saleItem
        required: true
        schema:
          $ref: '#/definitions/models.SaleItem'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SaleItem'
        "400":
          description: Неверный запрос или ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка сервера при обновлении позиции продажи
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Обновить позицию продажи
      tags:
      - sale-items
  /sales:
    get:
      description: Возвращает список всех продаж
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Sale'
            type: array
        "500":
          description: Ошибка сервера при получении продаж
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Получить все продажи
      tags:
      - sales
    post:
      consumes:
      - application/json
      description: Создает новую продажу
      parameters:
      - description: Данные продажи
        in: body
        name: sale
        required: true
        schema:
          $ref: '#/definitions/models.Sale'
      produces:
      - application/json
      responses:
        "200":
          description: Успешное создание
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка сервера при создании продажи
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Создать продажу
      tags:
      - sales
  /sales/{id}:
    delete:
      description: Удаляет продажу по ID
      parameters:
      - description: ID продажи
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Удаление успешно
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверный ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка сервера при удалении
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Удалить продажу
      tags:
      - sales
    get:
      description: Возвращает продажу по идентификатору
      parameters:
      - description: ID продажи
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Sale'
        "400":
          description: Неверный ID продажи
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка сервера при получении продажи
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Получить продажу по ID
      tags:
      - sales
    put:
      consumes:
      - application/json
      description: Обновляет данные продажи по ID
      parameters:
      - description: ID продажи
        in: path
        name: id
        required: true
        type: integer
      - description: Обновленные данные продажи
        in: body
        name: sale
        required: true
        schema:
          $ref: '#/definitions/models.Sale'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Sale'
        "400":
          description: Неверный запрос или ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка сервера при обновлении
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Обновить продажу
      tags:
      - sales
...
swagger: "2.0"